Remove unneeded imports and make code easier to read Fix potential infinite loop as moxie class with stills access
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@3915 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/MoonPhaseDatabase.java
@@ -209,9 +209,9 @@ public class MoonPhaseDatabase extends StaticEntity
 		// Adjust the new year by the appropriate
 		// number of days.
 
-		NEWYEAR += ((long)phaseError) * 86400000L;
-		BOUNDARY += ((long)phaseError) * 86400000L;
-		COLLISION += ((long)phaseError) * 86400000L;
+		NEWYEAR += (phaseError) * 86400000L;
+		BOUNDARY += (phaseError) * 86400000L;
+		COLLISION += (phaseError) * 86400000L;
 	}
 
 	public static final int getRonaldPhase()
